Option to have rest blocks merged into master.

+++ stimulus_presentation/Code/prfStim_CreateConditions.py
@@ -4,9 +4,14 @@
 from __future__ import division  # so that 1/3=0.333 instead of 1/3=0
 import numpy as np
 import os
+import errno
 import pickle
 
-# %% set parameters
+# %% Set parameters
+
+# set folder for stimulation condition
+OutFolderPath = '/home/faruk/Documents/test/Conditions'
+OutFileName = 'Conditions_run01.pickle'
 
 # set the TR
 TR = 1.20
@@ -32,9 +37,7 @@ if NrNullFixBetween > 0:
     # how many TRs should one rest preiod be? (default 1)
     NrOfTrPerNull = 3
 
-NrOfVolsExclNull = (NrOfBlocks * NrOfOrientation * NrOfSteps)
-
-# %% create arrays for position and orientation
+# %% Create arrays for position and orientation
 aryOri = np.empty(0)
 aryPos = np.empty(0)
 for indBlock in np.arange(NrOfBlocks):
@@ -54,20 +57,24 @@ conditions = np.vstack((aryPos, aryOri)).T
 
 if NrNullFixBetween > 0:
     while lgcRep:
-        NullPos = np.random.choice(np.arange(1, NrOfVolsExclNull),
+        NullPos = np.random.choice(np.arange(1, len(conditions)),
                                    NrNullFixBetween, replace=False)
-        lgcRep = np.greater(np.sum(np.diff(np.sort(NullPos)) == 1), 0)
+        NullPos = np.sort(NullPos)
+        lgcRep = np.greater(np.sum(np.diff(NullPos) == 1), 0)
 
-# insert null trials in between
-conditions = np.insert(conditions, NullPos, np.array([0, 0]),
-                       axis=0)
+    # create null trials in between to be inserted into conditions
+    NullTrialsInBetween = np.zeros(NrOfTrPerNull)[:, None]
+    # insert null trials in between
+    for i, idx in enumerate(NullPos):
+        idx = idx + (i*NrOfTrPerNull)  # adjustment to consider prev. iterations
+        conditions = np.insert(conditions, idx, NullTrialsInBetween, axis=0)
 
 # add fixation blocks in beginning and end
 conditions = np.vstack((np.zeros((NrNullFixStart, 2)),
                         conditions,
                         np.zeros((NrNullFixEnd, 2))))
 
-# %% prepare target arrays
+# %% Prepare target arrays
 NrOfTargets = int(len(conditions)/10)
 targets = np.zeros(len(conditions))
 lgcRep = True
@@ -80,7 +87,7 @@ targets[targetPos] = 1
 assert NrOfTargets == np.sum(targets)
 targets = targets.astype(bool)
 
-# %% prepare random target onset delay
+# %% Prepare random target onset delay
 BlockOnsetinSec = np.arange(len(conditions)) * TR
 TargetOnsetinSec = BlockOnsetinSec[targets]
 TargetOnsetDelayinSec = np.random.uniform(0.1,
@@ -88,7 +95,7 @@ TargetOnsetDelayinSec = np.random.uniform(0.1,
                                           size=NrOfTargets)
 TargetOnsetinSec = TargetOnsetinSec + TargetOnsetDelayinSec
 
-# %% create dictionary for saving to pickle
+# %% Create dictionary for saving to pickle
 array_run1 = {'Conditions': conditions,
               'TargetOnsetinSec': TargetOnsetinSec,
               'TR': TR,
@@ -97,9 +104,13 @@ array_run1 = {'Conditions': conditions,
               'NrOfVols': len(conditions),
               }
 
-# %% save dictionary to pickle
-folderpath = '/media/sf_D_DRIVE/ParamContrast/pRFStimuli/prfStim/prfStim/Conditions'
-filename1 = os.path.join(folderpath, 'Conditions_run01.pickle')
+# %% Save dictionary to pickle
+OutFile = os.path.join(OutFolderPath, OutFileName)
 
-with open(filename1, 'wb') as handle:
-    pickle.dump(array_run1, handle)
+try:
+    os.makedirs(OutFolderPath)
+except OSError as e:
+    if e.errno != errno.EEXIST:
+        raise
+    with open(OutFile, 'wb') as handle:
+        pickle.dump(array_run1, handle)
